+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Support for review frequency in statistics page.
+
+### Changed
+- Persist study page filters in URL.
+
 ## [0.8.0] - 2020-08-03
 ### Added
 - Labels for decks and models sections of homepage.